Update monte_carlo.py
Thank you for your edits, the code is much faster now, great!
I also added the calculation of the electricity costs for the heat pump, which was missing. 

Since I have to do this calculation for 60 different geothermal system, is there a way to loop this calculation over all these systems?